Regression in text hotspots with "Use custom styles"
In hotspots where activation on "click / tap / always" is not set, a centered popup on mouse hover would close prematurely as soon as the mouse is not on the indicator. The fixed issue only applies to hotspots with "Use custom styles" checked under Text Content tab in SpotEditor and where popup is configured to open on top of the hotspot indicator.